% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMRAS.R
\docType{package}
\name{IMRAS}
\alias{IMRAS}
\alias{IMRAS-package}
\alias{IMRAS-package}
\title{IMRAS}
\description{
Raw read counts and TPM values for IMRAS RNASeq assay.
}
\details{
IMRAS
A data package for IMRAS.

\code{ExpressionSet}s count_eset (raw counts) and tpm_eset (tpm values) generated with STAR and RSEM pipeline.
}
\seealso{
\link{count_eset}
\link{tpm_eset}
}
