% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMRAS.R
\docType{data}
\name{tpm_eset}
\alias{tpm_eset}
\title{TPM values}
\format{a \code{ExpressionSet} containing the following fields:
\describe{
 \item{\code{assayData}:}{Matrix of class \code{numeric} containing TPM values generated from STAR+RSEM in R package RNASeqPipelineR. Each row is a gene and each column is a sample. Row names are gene symbols.}
   \item{\code{phenoData}:}{Matrix of mixed class, containing phenotype
(experimental design) data for each sample. Each row matches the corresponding column in assayData. Each column is an experimental factor}
   \item{\code{featureData}:}{Matrix of class \code{character} defining the gene annoations in assayData. Each row maps to the corresponding row in assayData.}
}}
\source{
The data comes from the IMRAS project.
}
\description{
Transcript per million (TPM) values of IMRAS RNASeq data.
}
\seealso{
\link{IMRAS}
\link{count_eset}
}
